make most funcs private
